Created Components folder & added SignInScreen component ,also created Navigators folder & added Stack component for the stackNavigation
