+++ Components/SignInScreen.js
@@ -1,0 +1,25 @@
+import React from 'react';
+
+import {View, Text, StyleSheet} from 'react-native';
+
+const SignInScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.text}>Hello World</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  text: {
+    fontSize: 20,
+    fontWeight: '700',
+  },
+});
+
+export default SignInScreen;
